passwordViewModel , NameViewModel 추가
