+++ SignUp/SignUp/ViewModel/NameViewModel.swift
@@ -1,0 +1,22 @@
+//
+//  NameViewModel.swift
+//  SignUp
+//
+//  Created by HOONHA CHOI on 2021/03/31.
+//
+
+import Foundation
+import Combine
+
+class NameViewModel {
+    
+    @Published var nameText = ""
+    
+    var isNameValid : AnyPublisher<NameState, Never> {
+        $nameText
+            .dropFirst()
+            .map { $0.isEmpty ? NameState.empty : NameState.valid }
+            .eraseToAnyPublisher()
+    }
+    
+}
